Add files via upload
Fixed bug where it would ignore the MIN or the MAX if only one of the two was specified. Also added optional features to plot the KDE (off by default). 

Also, I should point out that I noticed that the final "bandwidth_cdf=np.sqrt(t_cdf)*R" is not used in any way, so perhaps it should be removed?